Extracting features for training and training the model finished. Let's see if
there is any improvement
