+++ src/main/scala/com/github/pedrovgs/roma/SparkApp.scala
@@ -14,7 +14,7 @@ private[roma] trait SparkApp extends App {
 
   private lazy val conf: SparkConf =
     new SparkConf()
-      .set("spark.kryoserializer.buffer.mb", "256")
+      .set("spark.kryoserializer.buffer", "256MB")
       .set("spark.kryoserializer.buffer.max", "512")
       .set("spark.serializer", "org.apache.spark.serializer.KryoSerializer")
