+++ src/main/scala/com/github/pedrovgs/roma/machinelearning/Config.scala
@@ -1,0 +1,11 @@
+package com.github.pedrovgs.roma.machinelearning
+
+private[machinelearning] object Config {
+  val numberOfIterations: Int = 100
+  val relevantWrods = "relevant"
+  val tweetWordsColumnName = "words"
+  val tweetContentColumnName = "tweetContent"
+  val rawFeaturesColumnName = "rawFeatures"
+  val featuresColumnName = "features"
+  val labelColumnName = "label"
+}
