+++ src/main/scala/com/github/pedrovgs/roma/machinelearning/TweetClassifier.scala
@@ -1,0 +1,5 @@
+package com.github.pedrovgs.roma.machinelearning
+
+object TweetClassifier {
+
+}
